[DHT] Refactor session code into the main class and thread listen
