+++ src/dht/sessionevents.cpp
@@ -4,7 +4,6 @@
 
 #include "dht/sessionevents.h"
 
-#include "dht/session.h"
 #include "sync.h" // for LOCK and CCriticalSection
 #include "util.h" // for LogPrintf
 #include "utiltime.h" // for GetTimeMillis
@@ -22,16 +21,9 @@
 
 using namespace libtorrent;
 
-typedef std::pair<int64_t, CEvent> EventPair;
-typedef std::multimap<int, EventPair> EventTypeMap;
-typedef std::map<std::string, CMutableGetEvent> DHTGetEventMap;
 
-static CCriticalSection cs_EventMap;
-static CCriticalSection cs_DHTGetEventMap;
 
-static bool fShutdown;
-static EventTypeMap m_EventTypeMap;
-static DHTGetEventMap m_DHTGetEventMap;
+
 
 CEvent::CEvent(const std::string& _message, const int _type, const uint32_t _category, const std::string& _what)
 {
@@ -94,165 +86,4 @@ CPutRequest::CPutRequest(const CKeyEd25519& _key, const std::string& _salt, cons
     sequence = _sequence;
     value = _value;
     timestamp = GetTimeMillis();
-}
-
-void AddToDHTGetEventMap(const std::string& infoHash, const CMutableGetEvent& event)
-{
-    LOCK(cs_DHTGetEventMap);
-    if (m_DHTGetEventMap.find(infoHash) == m_DHTGetEventMap.end()) {
-        // event not found. Add a new entry to DHT event map
-        LogPrint("dht", "AddToDHTGetEventMap Not found -- infohash = %s\n", infoHash);
-        m_DHTGetEventMap.insert(std::make_pair(infoHash, event));
-    }
-    else {
-        // event found. Update entry in DHT event map
-        LogPrint("dht", "AddToDHTGetEventMap Found -- Updateinfohash = %s\n", infoHash);
-        m_DHTGetEventMap[infoHash] = event;
-    }
-}
-
-static void AddToEventMap(const int type, const CEvent& event)
-{
-    LOCK(cs_EventMap);
-    m_EventTypeMap.insert(std::make_pair(type, std::make_pair(event.Timestamp(), event)));
-}
-
-static void DHTEventListener(session* dhtSession)
-{
-    SetThreadPriority(THREAD_PRIORITY_LOWEST);
-    RenameThread("dht-events");
-    unsigned int counter = 0;
-    while(!fShutdown)
-    {
-        if (!dhtSession->is_dht_running()) {
-            LogPrint("dht", "%s -- DHT is not running yet\n", __func__);
-            MilliSleep(2000);
-            continue;
-        }
-
-        dhtSession->wait_for_alert(seconds(1));
-        std::vector<alert*> alerts;
-        dhtSession->pop_alerts(&alerts);
-        for (std::vector<alert*>::iterator iAlert = alerts.begin(), end(alerts.end()); iAlert != end; ++iAlert) {
-            if ((*iAlert) == nullptr)
-                continue;
-
-            const uint32_t iAlertCategory = (*iAlert)->category();
-            const std::string strAlertMessage = (*iAlert)->message();
-            const int iAlertType = (*iAlert)->type();
-            const std::string strAlertTypeName = alert_name(iAlertType);
-            if (iAlertType == DHT_GET_ALERT_TYPE_CODE || iAlertType == DHT_PUT_ALERT_TYPE_CODE) {
-                if (iAlertType == DHT_GET_ALERT_TYPE_CODE) {
-                    // DHT Get Mutable Event
-                    dht_mutable_item_alert* pGet = alert_cast<dht_mutable_item_alert>((*iAlert));
-                    if (pGet == nullptr)
-                        continue;
-                    LogPrint("dht", "%s -- PubKey = %s, Salt = %s, Value = %s\nMessage = %s, Alert Type =%s, Alert Category = %u\n"
-                        , __func__, aux::to_hex(pGet->key), pGet->salt, pGet->item.to_string(), strAlertMessage, strAlertTypeName, iAlertCategory);
-
-                    if (pGet->item.to_string() != "<uninitialized>") {
-                        const CMutableGetEvent event(strAlertMessage, iAlertType, iAlertCategory, strAlertTypeName, 
-                          aux::to_hex(pGet->key), pGet->salt, pGet->seq, pGet->item.to_string(), aux::to_hex(pGet->signature), pGet->authoritative);
-
-                        std::string infoHash = GetInfoHash(event.PublicKey(), event.Salt());
-                        AddToDHTGetEventMap(infoHash, event);
-                    }
-                }
-            }
-            else if (iAlertType == DHT_STATS_ALERT_TYPE_CODE) {
-                // TODO (dht): handle stats 
-                //dht_stats_alert* pAlert = alert_cast<dht_stats_alert>((*iAlert));
-                LogPrintf("%s -- DHT Alert Message: AlertType = %s\n", __func__, strAlertTypeName); 
-            }
-            else {
-                const CEvent event(strAlertMessage, iAlertType, iAlertCategory, strAlertTypeName);
-                AddToEventMap(iAlertType, event);
-            }
-        }
-        if (fShutdown)
-            return;
-        
-        counter++;
-        if (counter % 60 == 0) {
-            LogPrint("dht", "DHTEventListener -- Before CleanUpEventMap. counter = %u\n", counter);
-            CleanUpEventMap(300000);
-        }
-    }
-}
-
-void CleanUpEventMap(const uint32_t timeout)
-{
-    unsigned int deleted = 0;
-    unsigned int counter = 0;
-    int64_t iTime = GetTimeMillis();
-    LOCK(cs_EventMap);
-    for (auto it = m_EventTypeMap.begin(); it != m_EventTypeMap.end(); ) {
-        CEvent event = it->second.second;
-        if ((iTime - event.Timestamp()) > timeout) {
-            it = m_EventTypeMap.erase(it);
-            deleted++;
-        }
-        else {
-            ++it;
-        }
-        counter++;
-    }
-    LogPrint("dht", "DHTEventListener -- CleanUpEventMap. deleted = %u, count = %u\n", deleted, counter);
-}
-
-void StopEventListener()
-{
-	fShutdown = true;
-    LogPrint("dht", "DHTEventListener -- stopping.\n");
-    MilliSleep(2100);
-}
-
-void StartEventListener(session* dhtSession)
-{
-    LogPrint("dht", "StartEventListener -- start\n");
-    fShutdown = false;
-    DHTEventListener(dhtSession);
-}
-
-bool GetLastTypeEvent(const int& type, const int64_t& startTime, std::vector<CEvent>& events)
-{
-    //LOCK(cs_EventMap);
-    LogPrint("dht", "GetLastTypeEvent -- m_EventTypeMap.size = %u, type = %u.\n", m_EventTypeMap.size(), type);
-    std::multimap<int, EventPair>::iterator iEvents = m_EventTypeMap.find(type);
-    while (iEvents != m_EventTypeMap.end()) {
-        if (iEvents->second.first >= startTime) {
-            events.push_back(iEvents->second.second);
-        }
-        iEvents++;
-    }
-    LogPrint("dht", "GetLastTypeEvent -- events.size() = %u\n", events.size());
-    return events.size() > 0;
-}
-
-bool FindDHTGetEvent(const std::string& infoHash, CMutableGetEvent& event)
-{
-    //LOCK(cs_DHTGetEventMap);
-    std::map<std::string, CMutableGetEvent>::iterator iMutableEvent = m_DHTGetEventMap.find(infoHash);
-    if (iMutableEvent != m_DHTGetEventMap.end()) {
-        // event found.
-        event = iMutableEvent->second;
-        return true;
-    }
-    return false;
-}
-
-bool RemoveDHTGetEvent(const std::string& infoHash)
-{
-    LOCK(cs_DHTGetEventMap);
-    m_DHTGetEventMap.erase(infoHash);
-    return true;
-}
-
-bool GetAllDHTGetEvents(std::vector<CMutableGetEvent>& vchGetEvents)
-{
-    //LOCK(cs_DHTGetEventMap);
-    for (std::map<std::string, CMutableGetEvent>::iterator it=m_DHTGetEventMap.begin(); it!=m_DHTGetEventMap.end(); ++it) {
-        vchGetEvents.push_back(it->second);
-    }
-    return true;
 }
